==> notebooks/2_model_creation.ipynb <==
import pandas as pd
import numpy as np
from sklearn.model_selection import train_test_split
from sklearn.metrics import mean_absolute_error, mean_absolute_percentage_error, accuracy_score
import joblib

RANDOM_SEED = 42

df = pd.read_parquet('../data/processed/1_train_processed.parquet')

df

X = df.drop(columns=['id', 'sales'])
y = df['sales']

X.shape, y.shape

X_train, X_test, y_train, y_test = train_test_split(X, y, test_size=0.2, random_state=RANDOM_SEED)

X_train.shape, X_test.shape, y_train.shape, y_test.shape

len(df[df.sales == 0]) / len(df)

%%time

from sklearn.ensemble import GradientBoostingClassifier

df_chunk = df.sample(10000, random_state=RANDOM_SEED)

df_chunk['is_zero'] = (df_chunk['sales'] == 0).astype(int)

X = df_chunk.drop(columns=['id', 'sales', 'is_zero'])
y = df_chunk['is_zero']

X_train, X_test, y_train, y_test = train_test_split(X, y, test_size=0.2, random_state=42)

clf = GradientBoostingClassifier()
clf.fit(X_train, y_train)
y_train_pred = clf.predict(X_train)
y_test_pred = clf.predict(X_test)

train_acc = accuracy_score(y_true=y_train, y_pred=y_train_pred)
test_acc = accuracy_score(y_true=y_test, y_pred=y_test_pred)

f"{train_acc = :.2f}, {test_acc = :.2f}"

from sklearn.ensemble import HistGradientBoostingRegressor

df_non_zero = df_chunk[df_chunk['sales'] > 0]

X_non_zero = df_non_zero.drop(columns=['sales', 'id', 'is_zero'])
y_non_zero = df_non_zero['sales']

X_train, X_test, y_train, y_test = train_test_split(X_non_zero, y_non_zero, test_size=0.2, random_state=42)

reg = HistGradientBoostingRegressor()
reg.fit(X_train, y_train)
y_pred = reg.predict(X_test)

mae = mean_absolute_error(y_test, y_pred)
mape = mean_absolute_percentage_error(y_test, y_pred)

f"{mae = :.2f}, {mape = :.2f}%"

df['is_zero'] = (df['sales'] == 0).astype('bool')

%%time

from sklearn.ensemble import GradientBoostingClassifier

X = df.drop(columns=['id', 'sales', 'is_zero'])
y = df['is_zero']

X_train, X_test, y_train, y_test = train_test_split(X, y, test_size=0.2, random_state=RANDOM_SEED)

clf = GradientBoostingClassifier()
clf.fit(X_train, y_train)
y_train_pred = clf.predict(X_train)
y_test_pred = clf.predict(X_test)

train_acc = accuracy_score(y_true=y_train, y_pred=y_train_pred)
test_acc = accuracy_score(y_true=y_test, y_pred=y_test_pred)

f"{train_acc = :.2f}, {test_acc = :.2f}"

%%time

from sklearn.ensemble import HistGradientBoostingRegressor

df_non_zero = df[df.is_zero == False]

X = df_non_zero.drop(columns=['sales', 'id', 'is_zero'])
y = df_non_zero['sales']

X_train, X_test, y_train, y_test = train_test_split(X, y, test_size=0.2, random_state=RANDOM_SEED)

hgbr = HistGradientBoostingRegressor()
hgbr.fit(X_train, y_train)
y_pred = hgbr.predict(X_test)

mae = mean_absolute_error(y_test, y_pred)
mape = mean_absolute_percentage_error(y_test, y_pred)

f"{mae = :.2f}, {mape = :.2f}%"

joblib.dump(clf, '../models/clf_1.pkl')
joblib.dump(hgbr, '../models/hgbr_1.pkl')

from xgboost import XGBClassifier

X = df.drop(columns=['id', 'sales', 'is_zero'])
y = df['is_zero']

X_train, X_test, y_train, y_test = train_test_split(X, y, test_size=0.2, random_state=RANDOM_SEED)

xgbc = XGBClassifier()
xgbc.fit(X_train, y_train)

y_train_pred = xgbc.predict(X_train)
y_test_pred = xgbc.predict(X_test)

train_acc = accuracy_score(y_true=y_train, y_pred=y_train_pred)
test_acc = accuracy_score(y_true=y_test, y_pred=y_test_pred)

print(f"{train_acc = :.2f}, {test_acc = :.2f}")

from xgboost import XGBRegressor

df_non_zero = df[df.is_zero == False]

X = df_non_zero.drop(columns=['sales', 'id', 'is_zero'])
y = df_non_zero['sales']

X_train, X_test, y_train, y_test = train_test_split(X, y, test_size=0.2, random_state=RANDOM_SEED)

xgbr = XGBRegressor()
xgbr.fit(X_train, y_train)

y_pred = xgbr.predict(X_test)

mae = mean_absolute_error(y_test, y_pred)
mape = mean_absolute_percentage_error(y_test, y_pred)

print(f"{mae = :.2f}, {mape = :.2f}%")

joblib.dump(xgbc, '../models/xgbc_1.pkl')
joblib.dump(xgbr, '../models/xgbr_1.pkl')
