==> notebooks/3_prediction_processing.ipynb <==
import pandas as pd
import joblib

df = pd.read_parquet('../data/processed/1_test_processed.parquet')

df

clf = joblib.load('../models/clf_1.pkl')
hgbr = joblib.load('../models/hgbr_1.pkl')

df_1 = df.copy()

X = df_1.drop(columns=['id'])

df_1['is_zero'] = clf.predict(X)

df_1

non_zero_rows = df_1[df_1['is_zero'] == False].copy()

non_zero_rows

y = hgbr.predict(non_zero_rows.drop(columns=['is_zero', 'id']))

df_1.loc[df_1['is_zero'] == False, 'sales'] = y
df_1.loc[df_1['is_zero'] == True, 'sales'] = 0

df_1

((len(df_1) - len(df_1[df_1.sales == 0])) / len(df_1)) * 100

df_train = pd.read_csv('../data/row/train.csv')

((len(df_train) - len(df_train[df_train.sales == 0])) / len(df_train)) * 100

df_1[df_1['id'].duplicated(keep=False)]

df_1[['id', 'sales']].to_csv('../data/processed/submission_1.csv', index=False)

xgbc = joblib.load('../models/xgbc_1.pkl')
xgbr = joblib.load('../models/xgbr_1.pkl')

df_2 = df.copy(deep=True)

X = df_2.drop(columns=['id'])

df_2['is_zero'] = xgbc.predict(X)

non_zero_rows = df_2[df_2['is_zero'] == False].copy()

y = xgbr.predict(non_zero_rows.drop(columns=['is_zero', 'id']))

df_2.loc[df_2['is_zero'] == False, 'sales'] = y
df_2.loc[df_2['is_zero'] == True, 'sales'] = 0

df_2

df_2.sales[df_2.sales < 0] = 0

df_2[df_2.sales < 0]

(len(df_2) - len(df_2[df_2.sales == 0])) / len(df_2) * 100

df_2[['id', 'sales']].to_csv('../data/processed/submission_2.csv', index=False)
