==> notebooks/4_NN_model_creation.ipynb <==
import pandas as pd
import torch
from torch import nn
from torch.utils.data import DataLoader, TensorDataset, random_split
from sklearn.model_selection import train_test_split

RANDOM_SEED = 42
random_generator = torch.Generator().manual_seed(RANDOM_SEED)

device = "mps" if torch.backends.mps.is_available() else "cpu"

device

df = pd.read_parquet('../data/processed/1_train_processed.parquet')

df

df['is_zero'] = (df.sales == 0).astype('float32')

df

df.shape

df.info()

# df = df.head(1_000_000)

df = df.astype('float32')

df.info()

X = df.drop(columns=['id', 'sales', 'is_zero'])
y_is_zero = df['is_zero']
y_sales = df['sales']

X.shape, y_is_zero.shape, y_sales.shape

# X_tensor = torch.tensor(X.to_numpy(), dtype=torch.float32, device=device)
# y_is_zero_tensor = torch.tensor(y_is_zero.to_numpy(), dtype=torch.float32, device=device)
# y_sales_tensor = torch.tensor(y_sales.to_numpy(), dtype=torch.float32, device=device)

# X_tensor.shape, y_is_zero_tensor.shape, y_sales_tensor.shape

class SalesPredictor(nn.Module):
    def __init__(self):
        super(SalesPredictor, self).__init__()
        self.shared_layers = nn.Sequential(
            nn.Linear(X.shape[1], 128),
            nn.ReLU(),
            nn.Dropout(p=0.3),
            nn.Linear(128, 64),
            nn.ReLU()
        )
        # self.classification_layers = nn.Sequential(
        #     nn.Linear(64, 32),
        #     nn.ReLU(),
        #     nn.Linear(32, 1),
        #     nn.Sigmoid()
        # )
        self.regression_layers = nn.Sequential(
            nn.Linear(64, 32),
            nn.ReLU(),
            nn.Linear(32, 1)
        )
    def forward(self, x):
        shared_layers = self.shared_layers(x)
        # is_zero = self.classification_layers(shared_layers)
        sales = self.regression_layers(shared_layers)
        return sales

X_train, X_test, y_is_zero_train, y_is_zero_test, y_sales_train, y_sales_test = train_test_split(
    X, y_is_zero, y_sales, test_size=0.2, random_state=RANDOM_SEED
)

X_train.shape, X_test.shape, y_is_zero_train.shape, y_is_zero_test.shape, y_sales_train.shape, y_sales_test.shape

X_train_tensor = torch.tensor(X_train.to_numpy(), dtype=torch.float32, device=device)
X_test_tensor = torch.tensor(X_test.to_numpy(), dtype=torch.float32, device=device)
y_is_zero_train_tensor = torch.tensor(y_is_zero_train.to_numpy(), dtype=torch.float32, device=device)
y_is_zero_test_tensor = torch.tensor(y_is_zero_test.to_numpy(), dtype=torch.float32, device=device)
y_sales_train_tensor = torch.tensor(y_sales_train.to_numpy(), dtype=torch.float32, device=device)
y_sales_test_tensor = torch.tensor(y_sales_test.to_numpy(), dtype=torch.float32, device=device)

print(f"""{X_train_tensor.shape = }, {X_test_tensor.shape = }, {y_is_zero_train_tensor.shape = }, 
      {y_is_zero_test_tensor.shape = }, {y_sales_train_tensor.shape = }, {y_sales_test_tensor.shape = }""")

# ds = TensorDataset(X_tensor, y_is_zero_tensor, y_sales_tensor)

# train_size = int(0.8 * len(ds))
# test_size = int(len(ds) - train_size)
# train_ds, test_ds = random_split(dataset=ds, lengths=[train_size, test_size], generator=random_generator)

# train_ds, test_ds

# BATCH_SIZE = 128

# train_dataloader = DataLoader(train_ds, batch_size=BATCH_SIZE, shuffle=True, generator=random_generator)
# test_dataloader = DataLoader(test_ds, batch_size=BATCH_SIZE, shuffle=False, generator=random_generator)

# f"number of batches = {X.shape[0] / BATCH_SIZE}"

torch.manual_seed(RANDOM_SEED)
model = SalesPredictor()

model.to(device)

model, model.state_dict(), next(model.parameters()).device

LEARNING_RATE = 0.01

loss_clf_fn = nn.BCELoss()
loss_reg_fn = nn.L1Loss()
optimizer = torch.optim.Adam(params=model.parameters(), lr=LEARNING_RATE)

def accuracy_fn(true_labels, pred_labels):
    correct = torch.eq(true_labels, pred_labels).sum().item()
    acc = (correct/len(pred_labels)) * 100
    return acc

# is_zero_pred, sales_pred = model(X_train_tensor)
# is_zero_pred = torch.round(is_zero_pred).squeeze()

# is_zero_pred.shape, y_is_zero_train_tensor.shape

y_sales_train_tensor.shape

# is_zero_pred, sales_pred = model(X_train_tensor)

# is_zero_pred[0]

# torch.round(is_zero_pred[0])

torch.manual_seed(RANDOM_SEED)
EPOCHS = 100

for epoch in range(EPOCHS):

    model.train()
    sales_pred = model(X_train_tensor)

    # is_zero_pred = is_zero_pred.squeeze()
    sales_pred = sales_pred.squeeze()
    # loss_clf = loss_clf_fn(is_zero_pred, y_is_zero_train_tensor)
    # acc = accuracy_fn(true_labels=y_is_zero_train_tensor, pred_labels=torch.round(is_zero_pred))

    # 'y_sales' where 'y_is_zero' == 0
    # 'sales_pred' with only the rows from 'y_sales'

    # mask = (y_is_zero_train_tensor == 0).cpu().numpy()

    # print(f"Shape of sales_pred: {sales_pred.shape}")
    # print(f"Shape of y_sales_train_tensor: {y_sales_train_tensor.shape}")
    # print(f"Shape of mask: {mask.shape}")

    # sales_pred = sales_pred[mask]
    # y_sales_train_tensor = y_sales_train_tensor[mask]

    loss = loss_reg_fn(sales_pred, y_sales_train_tensor)

    # loss = loss_clf + loss_reg

    optimizer.zero_grad()
    loss.backward()
    optimizer.step()

    # if (epoch + 1) % 10 == 0:
    #     print(f"{loss_clf = :.2f}, {acc = :.2f}, {loss_reg = :.2f}, {mape = :.2f}")

    if (epoch + 1) % 10 == 0:
        print(f"{loss = :.2f}")
