==> notebooks/5_NN_model_creation_2.ipynb <==
import pandas as pd
import torch
from torch import nn
from sklearn.model_selection import train_test_split
from torch.utils.data import TensorDataset, DataLoader

RANDOM_SEED = 42
random_generator = torch.Generator().manual_seed(RANDOM_SEED)

device = "mps" if torch.backends.mps.is_available() else "cpu"

device

df = pd.read_parquet('../data/processed/1_train_processed.parquet')

df

df.info()

df = df.astype('float32')

df.info()

X = df.drop(columns=['id', 'sales'])
y = df['sales']

X.shape, y.shape

X_tensor = torch.tensor(X.to_numpy(), dtype=torch.float32)
y_tensor = torch.tensor(y.to_numpy(), dtype=torch.float32)

X_tensor.shape, y_tensor.shape, X_tensor.device, y_tensor.device

X_train, X_test, y_train, y_test = train_test_split(X_tensor, y_tensor, test_size=0.2, random_state=RANDOM_SEED)

X_train.shape, X_test.shape, y_train.shape, y_test.shape

X_train.shape[1]

train_dataset = TensorDataset(X_train, y_train)
test_dataset = TensorDataset(X_test, y_test)

BATCH_SIZE = 1280

train_dataloader = DataLoader(dataset=train_dataset, batch_size=BATCH_SIZE, shuffle=True, generator=random_generator)
test_dataloader = DataLoader(dataset=test_dataset, batch_size=BATCH_SIZE, shuffle=False, generator=random_generator)

train_dataloader, test_dataloader

class SalesPredictor(nn.Module):
    def __init__(self):
        super(SalesPredictor, self).__init__()
        self.layer_stack = nn.Sequential(
            nn.Linear(X_train.shape[1], 128),
            nn.ReLU(),
            nn.Dropout(0.3),
            nn.Linear(128, 64),
            nn.ReLU(),
            nn.Dropout(0.3),
            nn.Linear(64, 32),
            nn.ReLU(),
            nn.Linear(32, 1)
        )
    def forward(self, x):
        return self.layer_stack(x)

torch.manual_seed(RANDOM_SEED)
model = SalesPredictor()
model, next(model.parameters()).device

LEARNING_RATE = 0.001

loss_fn = nn.L1Loss()
optimizer = torch.optim.Adam(params=model.parameters(), lr=LEARNING_RATE)

for X, y in train_dataloader:
    y_pred = model(X).squeeze()
    print(f"{y_pred.shape = }, {y.shape = }")
    break

torch.manual_seed(RANDOM_SEED)
EPOCHS = 40
model = model.to(device)

for epoch in range(EPOCHS):
    
    train_loss = 0  
    model.train()
    for X, y in train_dataloader:
        X, y = X.to(device), y.to(device)
        y_pred = model(X).squeeze()
        loss = loss_fn(y_pred, y)
        train_loss += loss
        optimizer.zero_grad()
        loss.backward()
        optimizer.step()
    train_loss /= len(train_dataloader)

    test_loss = 0
    model.eval()
    with torch.inference_mode():
        for X, y in test_dataloader:
            X, y = X.to(device), y.to(device)
            test_pred = model(X).squeeze()
            test_loss += loss_fn(test_pred, y)
        test_loss /= len(test_dataloader)

    if (epoch+1) % 1 == 0:
        print(f"Epoch = {epoch+1}/{EPOCHS}, {train_loss = :.2f}, {test_loss = :.2f}")

torch.save(obj=model.state_dict(), f='../models/sales_predictor_model.pth')

df_test = pd.read_parquet('../data/processed/1_test_processed.parquet').astype('float32')

df_test

df_test.info()

X_test_tensor = torch.tensor(df_test.drop(columns=['id']).to_numpy(), dtype=torch.float32)

X_test_tensor.shape, X_test_tensor.device

model.eval()
with torch.inference_mode():
    X_test_tensor = X_test_tensor.to(device)
    y_test_pred = model(X_test_tensor)

y_test_pred

y_test_pred = y_test_pred.cpu().numpy()

y_test_pred

df_test['sales'] = y_test_pred

df_test

df_test.sample(10, random_state=RANDOM_SEED)

df_test['sales'] = df_test['sales'].apply(lambda x: 0 if x < 0 else x)

df_test.sample(10, random_state=RANDOM_SEED)

df_test['sales'].min(), df_test['sales'].max()

df_test.info()

df_submission = df_test[['id', 'sales']].copy()

df_submission['sales'] = df_submission['sales'].astype('uint16')
df_submission['id'] = df_submission['id'].astype('uint32')

df_submission

df_submission.sample(10, random_state=RANDOM_SEED)

df_submission.info()

df_submission.to_csv('../data/processed/submission_3.csv', index=False)
